server: Enable sphinx usage
This enables sphinx usage as a router by passing bitcoin mainnet
parameters for its NewRouter function.

This is less than ideal, but the chaincfg.Params object in sphinx is
only used for creating a btcutil.Address object, so as long as all nodes
are using the same identification mechanism, this should be fine.

In the future this will need replacement for a decent solution.