lnd: Fix wallet init deadlock condition
This fixes a deadlock condition arising due to using the wallet to
derive addresses before the wallet is fully synced and initialized.

The main issue for this commit is that deriving a node's private/public
id keypair requires generating addresses (and their implicit key) in the
wallet, which can only be done after the wallet has fully synced and has
performed all initialization procedures.

This changes the order of initialization of the main lnd node, such that
no other actions are performed until the underlying wallet has fully
synced to the network.