discovery: Add ability to cancel bootstrapping
This adds the ability to cancel the peer bootstrapping procedure by
adding a context to the SampleNodeAddrs function of bootstrappers. If
this context is canceled, the associated bootstraper returns even
if it has not yet found the target number of remote peers.

Note that this follows a different convention than other services of the
node, which use a Start()/Stop() architecture for maintaining ongoing
processes, but is a compromise to provide cancellation services vs
rewriting the entire bootstrapping facility at this moment.

In the future, bootstrapping should follow the same conventions as other
services, even if internally it remains using a context.