rpcserver: Check MaxInboundAmt on AddInvoice
This adds a check during invoice addition to verify that the requested
invoice amount can be received after accounting for the available
maximum inbound amount for opened and live channels.

This improves the UX for creating payment requests by giving an early
alert to users that a given invoice cannot possibly be paid, given that
the node does not have an open channel with an online peer with enough
inbound capacity (after accounting for channel reserves) to receive it.

This behavior can be disabled by passing the IgnoreMaxInboundAmt flag on
AddInvoice requests.

Note that this check does not _ensure_ the payment will be completed; it
merely verifies that the most basic requirement for that exists within
directly connected channels. The capacity of the nodes may have changed
once the payer actually initiates the process to settle the invoice.

The code is acommpanied by integration tests exercising the expected
behavior.