rpcserver: Check MaxOutboundAmt before sendPayment
This adds a check before a node attempts a payment for the maximum
outbound amount available for that node, in the same vein as the test
added for the AddInvoice() call.

It allows the node to check whether the minimum conditions for
successfully performing a payment (namely: it has enough outbound
bandwidth in one of its opened and live channels).

This behavior can be disabled by specifying the IgnoreMaxOutboundAmt
flag on the rpc call.