lnwallet: Add WalletSyncer and RPCSyncer
This adds the WalletSyncer interface and its first implementation,
RPCSyncer. This allows package users to specify that a particular
DcrWallet instance should use an RPC backend full node for
synchronization and network services.

It also modifies the daemon init code to specify the WalletSyncer
instead of relying on it to instantiate a specific chain client, which
improves abstraction.