rpcserver: Allow listing invalid invoices
As pointed out in the previous commit, with the change of the bech32
prefix for testnet/simnet/regnet, invoices stored in the db became
invalid and would fail decoding, which completely prevets invoice
decoding.

This changes how errors are handled in the ListInvoices() rpc call, so
that invalid invoices can still be returned (with an error indication)
and the remaining invoices can be listed.