lnwallet: Reenable fee estimation
Fee estimation through dcrd was previously disabled due to rpcclient not
offering the required EstimateSmartFee function. This was solved
upstream, so now fee estimation via dcrd can be re-enabled.

This also updates the go.mod and go.sum files to point to the (as of yet
untagged) version with the appropriate function.