itest: Improve stability of switch_ test
This improves the stability of the testSwitchDeliveryOfflinePersistence
test by ensuring Alice and Bob nodes are connected for the final stage
of the test.

Previously, it could happen that they weren't connected before asserting
the HTLCs in the test were settled, so the final leg of settlement
(Alice -> Bob) would fail to complete and Alice would still see the
HTLCs as active.