multi: Do not reconnect to inbound peers
This changes the funding manager and server components to ensure
reconnect attempts are not done to peers whose only recorded address is
an inbound one.

Inbound peers (specially end-users) are unlikely to be reachable via a
future outbound connection, so it doesn't make sense to actively try to
connect to them.