rpcserver: Return paymenthash when payment errors
This changes the code for the sendPayment endpoint so that the original
payment hash of attempted payment is returned when the payment fails for
some reason.

This allows streaming clients to correctly identify _which_ previously
sent payment request failed by storing a map of payment hash to payment
request then finding out the corresponding original request in case an
error is returned on the receiving end of the payment stream.