itest: Switch to local defaultChanAmt
This switches the default channel capacity to use a local var instead of
the global maxDecredFundingAmt.

The maximum funding amount was increased in a previous commit which
causes the tests to fail to pass due to being larger than the coins sent
to the standard lntest harness nodes.

It also fixes tests affected by the change to the maximum funding
timeout.