lntest: Add delay when starting harness node
This adds a 3 second delay after starting the node before the
initialization of a new lntest harness node is considered complete.

This is less than ideal, but there's currently no way to ensure that all
init procedures of a node have been completed before issuing commands,
so this is a stop-gap solution to ensure no test commands are sent
before a node is potentially ready for it.