build: Switch to v0.1.0
This resets the minor version of the package to 1 (full version: 0.1.0)
to identify that the Decred port of lnd is on a different development
track.

It also updates the Makefile to correctly use the full package path when
adding the commit data, so that the binary produced by `make install`
has the commit tag info.