release: Fix dcrlncli location and destination dir
This fixes the name of the dcrlncli main package which was still
referring to the older lncli, causing the release script to fail.

It also modifies the release script to generate all artifacts in a
"releases" dir and properly adds that destination dir as a .gitignore so
that it's not erroneously added.