make: Update to build/build-itest dcrlnd/dcrlncli
This changes the make targets build and build-itest to generate the
binaries with the correct names. Previously they were still generated
with the old lnd/lncli names.

It also updates the .gitignore file to ignore the new binary names and
the harness code to run the correct integration test binary.