release: Make windows build again
This fixes the windows build which as broken by a previous lint commit.

The syscal.Stdin var is a handle when building for windows (vs an int on
other platforms) so it needs a typecast to be usable for reading
passwords.

The fix was made by adding files with appropriate build directives for
each possible platform, so that the typecast is only performed on the
windows platform and linting tools don't complain about it.