itest: Split revoked uncooperative close retrib
This splits the testRevokedCloseRetributionRemoteHodl into two tests and
fixes them accordingly.

The first test ensures that Carol (the party relaying an old, invalid
and revoked commitment state) does _not_ have a chance to broadcast
second-level HTLC-redeeming transactions. This ensures Dave's (the
breached counterparty) justice transaction redeems entirely from the
commitment transaction and is the only transaction in the mempool.

The second test ensures that Carol has enough time to broadcast
second-level HTLC-redeeming transactions before Dave detects the
breached state and sends a sweeping justice transaction that also
redeems from them (along with the funds still time-locked in the
commitment).

Previously these two situations were tested in a single test but
unreliably. It was a matter of timing during test execution which of the
two justice paths were executed and one of them (the second one) was
checking for the wrong number of transactions in the mempool which
caused transient failures in the CI server.

Splitting them in two different tests ensures both possible redeem paths
are tested and that they verify the exact conditions in the network are
the expected ones.

Additionally, checks were added to verify the balances of the affected
parties are consistent with the expected at the end of the redeeming
process.